Give more space to scrollbar
This commit adjust the width and height of Album and Folder, such that
the last item of each row is not stacked with FlatList's scrollbar,
making navigation with scrollbar easier. ...though current solution is
not responsive, as it will not work on screen with different dimension.
Its not like there will be somebody else using this app. ...right?